refactor: localize scene engine prompt helpers
