+++ frontend/src/features/scene-engine/components/prompt-input/useAutoExpandTextarea.ts
