+++ frontend/src/features/scene-engine/components/prompt-input/useVoiceRecorder.ts
